// client/src/main/react/api/sales/models/ProductDto.ts
/**
 * Copyright 2026 Martin Doychinov
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProductDto = {
  productId?: number;
  productVersion?: number;
  productRefId?: number;
  productBarcode?: string;
  productName?: string;
  productNote?: string;
  productUnits?: number;
  productMeasure?: string;
  productAvailable?: number;
  productCode?: string;
  productCode1?: string;
  productCode2?: string;
  productCode3?: string;
  productCode4?: string;
  productCode5?: string;
  productCode6?: string;
  productCode7?: string;
  productCode8?: string;
  productCode9?: string;
  productCy?: string;
  productPrice?: number;
  productPrice1?: number;
  productPrice2?: number;
  productPrice3?: number;
  productPrice4?: number;
  productPrice5?: number;
  productPrice6?: number;
  productPrice7?: number;
  productPrice8?: number;
  productPrice9?: number;
};